This compiles and runs without crashing, but I need to check that the coordinates match line by line the previous version of ELBA. doing that now

+++ include/KmerOps.hpp
@@ -10,7 +10,8 @@
 #define MAX_ALLTOALL_MEM (128 * 1024 * 1024)
 #endif
 
-typedef uint16_t PosInRead;
+// typedef uint16_t PosInRead;
+typedef uint32_t PosInRead;
 typedef uint64_t ReadId;
 
 typedef std::array<PosInRead, UPPER_KMER_FREQ> POSITIONS;
